requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
fastapi>=0.104.0
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx>=0.25.0
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
jiter==0.10.0
limits==5.2.0
mcp==1.9.1
numpy>=1.24.0
oauthlib==3.2.2
openai==1.82.0
openai-agents==0.0.16
opencv-python==4.9.0.80
opentelemetry-api==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-httpx==0.54b1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
passlib==1.7.4
Pillow==10.2.0
pinecone-client==3.0.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
pydantic_core==2.33.2
python-dotenv>=1.0.0
python-jose==3.3.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
slowapi>=0.1.8
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.36.3
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
tweepy==4.15.0
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn>=0.24.0
wrapt==1.17.2
zipp==3.22.0
pandas>=2.0.0
python-dateutil>=2.8.2